Pages/graph/model.py
from langchain_groq import ChatGroq
import os
from dotenv import load_dotenv

load_dotenv()



# Initialize the language model
llm = ChatGroq(
  model="llama-3.1-8b-instant",
  api_key=os.getenv("GROQ_API_KEY")
)

